deploy_temp/requirements-azure.txt
Django>=4.2.0,<4.3.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.5
Pillow>=9.5.0
whitenoise>=6.4.0
requests>=2.30.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
django-cors-headers>=4.3.1
django-allauth==0.57.0
social-auth-app-django>=5.4.0
gunicorn>=21.0.0
python-decouple>=3.8
django-storages>=1.13.2
pytz>=2023.3
tqdm>=4.65.0
django-tinymce==3.7.1
beautifulsoup4==4.12.3
PyPDF2==3.0.1
django-extensions==3.2.3 